+++ arkiv-mock/src/test/kotlin/no/nav/soknad/archiving/arkivmock/IntegrationTest.kt
@@ -49,7 +49,7 @@ class IntegrationTest {
 
 	private fun createRequestData(personId: String) =
 		ArkivData(
-			Bruker(personId, "FNR"), LocalDate.now().format(DateTimeFormatter.ISO_DATE_TIME), emptyList(),
+			Bruker(personId, "FNR"), LocalDateTime.now().format(DateTimeFormatter.ISO_DATE_TIME), emptyList(),
 			personId, "INNGAAENDE", "NAV_NO", tema, title
 		)
 }
