+++ end-to-end-tests/pom.xml
@@ -17,7 +17,6 @@
 		<kafka.version>2.5.0</kafka.version>
 		<testcontainers.version>1.14.3</testcontainers.version>
 		<postgresql.version>42.2.18</postgresql.version>
-		<coroutines.version>1.3.7</coroutines.version>
 		<okhttp.version>4.7.2</okhttp.version>
 		<jackson.version>2.11.0</jackson.version>
 		<junit-jupiter.version>5.6.2</junit-jupiter.version>
@@ -31,11 +30,6 @@
 	</properties>
 
 	<dependencies>
-		<dependency>
-			<groupId>org.jetbrains.kotlinx</groupId>
-			<artifactId>kotlinx-coroutines-core</artifactId>
-			<version>${coroutines.version}</version>
-		</dependency>
 		<dependency>
 			<groupId>org.slf4j</groupId>
 			<artifactId>slf4j-simple</artifactId>
