+++ end-to-end-tests/src/test/kotlin/no/nav/soknad/arkivering/arkiveringendtoendtests/verification/VerificationTaskManager.kt
@@ -1,4 +1,4 @@
-package no.nav.soknad.arkivering.arkiveringendtoendtests.locks
+package no.nav.soknad.arkivering.arkiveringendtoendtests.verification
 
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.runBlocking
